Refactor(be): Cursor 생성 로직 Service -> Repository Layer로 이동
