+++ backend/grabtable/src/main/java/edu/skku/grabtable/review/infrastructure/ReviewRepositoryCustom.java
@@ -1,11 +1,11 @@
 package edu.skku.grabtable.review.infrastructure;
 
+import edu.skku.grabtable.common.domain.response.SliceResponse;
 import edu.skku.grabtable.review.domain.response.ReviewResponse;
-import java.util.List;
 
 public interface ReviewRepositoryCustom {
 
-    List<ReviewResponse> findByUserIdBeforeCursor(Long userId, Long cursor, int size);
+    SliceResponse<ReviewResponse> findByUserIdBeforeCursor(Long userId, Long cursor, int size);
 
-    List<ReviewResponse> findByStoreIdBeforeCursor(Long storeId, Long cursor, int size);
+    SliceResponse<ReviewResponse> findByStoreIdBeforeCursor(Long storeId, Long cursor, int size);
 }
